Remove legacy rich text image styles
The `<div class="rich-text">` wrapper around rich text was disabled by default in Wagtail 2.10 (https://docs.wagtail.io/en/stable/releases/2.10.html#div-class-rich-text-wrappers-removed-from-rich-text) and we'd prefer not to enable the legacy handler, to encourage Wagtail developers to follow current best practice.

There is already a `width: 100%` rule in place for all img elements, and since an image with `width: 100%` but no `height: auto` would render with the wrong aspect ratio, it's reasonable to assume that no images on the site rely on `height: auto` being absent. Adding this to the `img` rule therefore makes the `rich-text` rule redundant.

The `width: 100%` also meant that the left/right-alignment options for rich text images had no effect - we add a `width: 50%` to those to make them meaningful again.

Finally, the responsive-object wrapper for media embeds was dropped in 2.9, so those rules are unused since then. Anyone who feels strongly about restoring responsive embeds is welcome to contribute a PR :-)